Move entity construction utilities into `AnEntityLite`
