+++ server/src/main/java/org/spine3/server/entity/Entity.java
@@ -30,8 +30,6 @@ import org.spine3.server.entity.status.EntityStatus;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nullable;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static org.spine3.protobuf.Timestamps.getCurrentTime;
@@ -84,57 +82,11 @@ public abstract class Entity<I, S extends Message>
      * the call to {@link #getDefaultState()} relies on completed initialization
      * of the instance.
      */
-    void init() {
+    @Override
+    protected void init() {
+        super.init();
         setState(getDefaultState(), 0, getCurrentTime());
         this.status = EntityStatus.getDefaultInstance();
-    }
-
-    /**
-     * Obtains constructor for the passed entity class.
-     *
-     * <p>The entity class must have a constructor with the single parameter of type defined by
-     * generic type {@code <I>}.
-     *
-     * @param entityClass the entity class
-     * @param idClass the class of entity identifiers
-     * @param <E> the entity type
-     * @param <I> the ID type
-     * @return the constructor
-     * @throws IllegalStateException if the entity class does not have the required constructor
-     */
-    static <E extends Entity<I, ?>, I> Constructor<E> getConstructor(Class<E> entityClass, Class<I> idClass) {
-        try {
-            final Constructor<E> result = entityClass.getDeclaredConstructor(idClass);
-            result.setAccessible(true);
-            return result;
-        } catch (NoSuchMethodException ignored) {
-            throw noSuchConstructor(entityClass.getName(), idClass.getName());
-        }
-    }
-
-    private static IllegalStateException noSuchConstructor(String entityClass, String idClass) {
-        final String errMsg = String.format("%s class must declare a constructor with a single %s ID parameter.",
-                                            entityClass, idClass);
-        return new IllegalStateException(new NoSuchMethodException(errMsg));
-    }
-
-    /**
-     * Creates new entity and sets it to the default state.
-     *
-     * @param constructor the constructor to use
-     * @param id the ID of the entity
-     * @param <I> the type of entity IDs
-     * @param <E> the type of the entity
-     * @return new entity
-     */
-    static <I, E extends Entity<I, ?>> E createEntity(Constructor<E> constructor, I id) {
-        try {
-            final E result = constructor.newInstance(id);
-            result.init();
-            return result;
-        } catch (InvocationTargetException | InstantiationException | IllegalAccessException e) {
-            throw new IllegalStateException(e);
-        }
     }
 
     /**
