+++ server/src/test/java/org/spine3/server/entity/EntityShould.java
@@ -286,8 +286,8 @@ public class EntityShould {
 
     @Test
     public void obtain_entity_constructor_by_class_and_ID_class() {
-        final Constructor<BareBonesEntity> ctor = Entity.getConstructor(BareBonesEntity.class,
-                                                                        Long.class);
+        final Constructor<BareBonesEntity> ctor = AnEntityLite.getConstructor(BareBonesEntity.class,
+                                                                              Long.class);
 
         assertNotNull(ctor);
     }
@@ -298,8 +298,8 @@ public class EntityShould {
         final Timestamp before = Timestamps.secondsAgo(1);
 
         // Create and init the entity.
-        final Constructor<BareBonesEntity> ctor = Entity.getConstructor(BareBonesEntity.class, Long.class);
-        final Entity<Long, StringValue> entity = Entity.createEntity(ctor, id);
+        final Constructor<BareBonesEntity> ctor = AnEntityLite.getConstructor(BareBonesEntity.class, Long.class);
+        final Entity<Long, StringValue> entity = AnEntityLite.createEntity(ctor, id);
 
         final Timestamp after = Timestamps.getCurrentTime();
